Clean explorer brand details
