+++ explorer/src/components/Layout.tsx
@@ -3,10 +3,9 @@ import Footer from "./Footer";
 import NavBar from "./Navbar";
 
 const Layout: React.FC = ({ children }) => (
-  <main style={{overflow:'hidden'}}>
+  <main style={{ overflow: "hidden" }}>
     <NavBar />
     {children}
-    <Footer />
   </main>
 );
