+++ explorer/src/pages/index.tsx
@@ -1,230 +1,119 @@
-import ArrowForward from "@mui/icons-material/ArrowForward";
-import { Box, Button, Typography } from "@mui/material";
-import { Link as RouterLink, PageProps, graphql, useStaticQuery } from "gatsby";
-import React, { useEffect, useState } from "react";
-import GridWithCards from "../components/GridWithCards";
+import { PageProps } from "gatsby";
+import { Box, } from "@mui/material";
+import React, { useEffect } from "react";
+import ExplorerSearch from "../components/ExplorerSearch/ExplorerSearch"
+import ExplorerStats from "../components/ExplorerStats/ExplorerStats";
 import HeroText from "../components/HeroText";
 import Layout from "../components/Layout";
+import NetworkSelect from "../components/NetworkSelect";
+import shape1 from "../images/index/shape2.svg";
 import { SEO } from "../components/SEO";
-import { portal as portalUrl } from "../utils/urls";
+import shapes from "../images/shape.png";
 
+import { paralaxGsap, animateSwirl } from "../utils/animations";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
 
-import apps from "../images/index/apps2.png";
-import blob from "../images/index/blob.svg";
-import cross from "../images/index/cross.svg";
-import cube from "../images/index/cube.svg";
-import portal from "../images/index/portal.png";
-import protocols from "../images/index/protocol3.png";
-import shape1 from "../images/index/shape1.svg";
-import shape from "../images/shape.png";
-import shape2 from "../images/shape2.png";
-import { toTotals } from "../components/ExplorerStats/ExplorerStats";
-import { amountFormatter } from "../utils/explorer";
-import { paralaxGsap, fadeInGsap, animateSwirl } from "../utils/animations";
-import { useNetworkContext } from "../contexts/NetworkContext";
-import { chainEnums } from "../utils/consts";
-
-const featuredNumber = {
-  fontSize: 42,
-  fontFamily: "Switzer",
-  fontWeight: "bold",
-};
-
-function toNotionalTVL(response: any): number | undefined {
-  let tvl: number = 0;
-  try {
-    for (const key in response.notionalTVL) {
-      const amount = parseFloat(response.notionalTVL[key]);
-      tvl += amount;
-    }
-  } catch (error) {
-    console.log(`failed to get notional tvl, error: ${error}`);
-    return undefined;
-  }
-  return tvl;
+interface ExplorerQueryValues {
+  emitterChain: number;
+  emitterAddress: string;
+  targetChain: number;
+  sequence: string;
+  txId: string;
 }
 
-const IndexPage = ({ location }: PageProps) => {
-  const { activeNetwork } = useNetworkContext();
-  const { site } = useStaticQuery<IndexQueryType>(IndexStaticQuery);
-  const [tvl, setTvl] = useState<number | undefined>(undefined);
-  const [messageTotal, setMessageTotal] = useState<number | undefined>(
-    undefined
-  );
+const ExplorerPage = ({ location }: PageProps) => {
+  const [emitterChain, setEmitterChain] =
+    React.useState<ExplorerQueryValues["emitterChain"]>();
+  const [emitterAddress, setEmitterAddress] =
+    React.useState<ExplorerQueryValues["emitterAddress"]>();
+  const [targetChain, setTargetChain] =
+    React.useState<ExplorerQueryValues["targetChain"]>();
+  const [sequence, setSequence] =
+    React.useState<ExplorerQueryValues["sequence"]>();
+  const [txId, setTxId] = React.useState<ExplorerQueryValues["txId"]>();
+  const [doneReadingQueryParams, setDoneReadingQueryParams] =
+    React.useState<boolean>(false);
 
-  const controller = new AbortController();
-  const { signal } = controller;
+  React.useEffect(() => {
+    if (location.search) {
+      // take searchparams from the URL and set the values in the form
+      const searchParams = new URLSearchParams(location.search);
 
-  const logo = {
-    "@type": "ImageObject",
-    url: `${site.siteMetadata.siteUrl}/logo-and-name-stacked.png`,
-    height: "146",
-    width: "146",
-  };
-  const structuredData = {
-    "@context": "https://schema.org",
-    "@type": "Organization",
-    "@id": "https://wormholenetwork.com#organization",
-    mainEntityOfPage: "https://wormholenetwork.com#organization",
-    url: "https://wormholenetwork.com",
-    name: "Wormhole",
-    sameAs: [
-      "https://github.com/certusone/wormhole",
-      "https://t.me/wormholecrypto",
-      "https://twitter.com/wormholecrypto",
-      "https://wormholebridge.com",
-      "https://wormholecrypto.medium.com",
-      "https://discord.gg/wormholecrypto",
-    ],
-    alternateName: [
-      "wormhole network",
-      "wormhole protocol",
-      "wormhole bridge",
-      "wormhole crypto",
-      "certus one wormhole",
-      "solana wormhole",
-      "SOL wormhole",
-      "terra wormhole",
-      "LUNA wormhole",
-      "ethereum wormhole",
-      "ETH wormhole",
-      "binance wormhole",
-      "BSC wormhole",
-      "oasis wormhole",
-      "ROSE wormhole",
-      "avalanche wormhole",
-      "AVAX wormhole",
-    ],
-    description: "A cross-chain messaging protocol.",
-    image: logo,
-    logo: logo,
-  };
+      const emitterChainParam = searchParams.get("emitterChain");
+      const address = searchParams.get("emitterAddress");
+      const targetChainParam = searchParams.get("targetChain")
+      const seq = searchParams.get("sequence");
+      const tx = searchParams.get("txId");
+
+      // if the search params are different form values, update state
+      if (Number(emitterChainParam) !== emitterChain) {
+        setEmitterChain(Number(emitterChainParam) || undefined);
+      }
+      if (address !== emitterAddress) {
+        setEmitterAddress(address || undefined);
+      }
+      if (Number(targetChainParam) !== targetChain) {
+        setTargetChain(Number(targetChainParam) || undefined);
+      }
+      if (seq !== sequence) {
+        setSequence(seq || undefined);
+      }
+      if (tx !== txId) {
+        setTxId(tx || undefined);
+      }
+    } else {
+      // clear state
+      setEmitterChain(undefined);
+      setEmitterAddress(undefined);
+      setTargetChain(undefined)
+      setSequence(undefined);
+      setTxId(undefined);
+    }
+    // be explicit about when it is ok to render
+    setDoneReadingQueryParams(true);
+  }, [location.search]);
+
 
   const headerImage = React.useRef<HTMLCanvasElement>(null);
-  const gradient1 = React.useRef<HTMLCanvasElement>(null);
-  const gradient2 = React.useRef<HTMLCanvasElement>(null);
-
-  const shapeLeft = React.useRef<HTMLCanvasElement>(null);
+  const gradient_1 = React.useRef<HTMLCanvasElement>(null);
   const shapeRight = React.useRef<HTMLCanvasElement>(null);
-  const row1 = React.useRef<HTMLCanvasElement>(null);
-  const row2 = React.useRef<HTMLCanvasElement>(null);
-  const row3 = React.useRef<HTMLCanvasElement>(null);
-  const row4 = React.useRef<HTMLCanvasElement>(null);
-  const row5 = React.useRef<HTMLCanvasElement>(null);
-
-  function fetchStats() {
-    const tvlUrl = `${activeNetwork.endpoints.backendUrl}api/stats/notionaltvl`;
-    const messagesUrl = `${activeNetwork.endpoints.backendUrl}api/stats/totalmessages`;
-
-    fetch(tvlUrl, { signal })
-      .then((res) => {
-        if (res.ok) return res.json();
-      })
-      .then(
-        (result) => {
-          setTvl(toNotionalTVL(result));
-        },
-        (error) => {
-          if (error.name !== "AbortError")
-            console.error("failed fetching notional TVL. error: ", error);
-        }
-      );
-    fetch(messagesUrl, { signal })
-      .then((res) => {
-        if (res.ok) return res.json();
-      })
-      .then(
-        (result) => {
-          const totals = toTotals(result);
-          setMessageTotal(totals.TotalCount["*"]);
-        },
-        (error) => {
-          if (error.name !== "AbortError")
-            console.error("failed fetching totals. error: ", error);
-        }
-      );
-  }
 
   useEffect(() => {
-    fetchStats(); // fetchStats on first load
 
     gsap.registerPlugin(ScrollTrigger);
-
-    // let icons = row5.current?.children[0];
-
-    var viewportWidth =
-      window.innerWidth || document.documentElement.clientWidth;
+    var viewportWidth = window.innerWidth || document.documentElement.clientWidth;
     if (viewportWidth > 992) {
-      gsap.utils.toArray(row5.current).forEach((section: any, i) => {
-        const img1 =
-          section.children[0].children[0].children[0].children[0].children[0];
-        const img2 =
-          section.children[0].children[1].children[0].children[0].children[0];
-        const img3 =
-          section.children[0].children[2].children[0].children[0].children[0];
-
-        gsap.from(img1, {
-          rotation: 30,
-          duration: 2,
-          ease: "Power3.easeOut",
-          scrollTrigger: {
-            trigger: img1,
-          },
-        });
-
-        gsap.from(img2, {
-          rotation: 30,
-          duration: 3,
-          ease: "Power3.easeOut",
-          scrollTrigger: {
-            trigger: img2,
-          },
-        });
-
-        gsap.from(img3, {
-          rotation: -20,
-          duration: 3,
-          ease: "Power3.easeOut",
-          scrollTrigger: {
-            trigger: img3,
-          },
-        });
-      });
-
       animateSwirl(headerImage);
-      paralaxGsap(gradient1, 1000, "-50% 100%");
-      paralaxGsap(gradient2, 1000, "-50% 100%");
-      paralaxGsap(shapeLeft, 200, "20% 100%");
+      paralaxGsap(gradient_1, 1000, "-50% 100%");
       paralaxGsap(shapeRight, 200, "20% 100%");
-
-      fadeInGsap(row1);
-      fadeInGsap(row2);
-      fadeInGsap(row3);
-      fadeInGsap(row4);
-      fadeInGsap(row5);
     }
-
-    return function cleanup() {
-      // abort any in-flight requests
-      controller.abort();
-    };
-  }, []);
+    
+  }, [])
 
   return (
     <Layout>
       <SEO
-        // use default title for index
-        description="The best of blockchains. Move information and value anywhere."
+        title="Explorer"
+        description="Explore real-time movement of information and value around the Alephium bridge."
         pathname={location.pathname}
-      >
-        <script type="application/ld+json">
-          {JSON.stringify(structuredData, undefined, 4)}
-        </script>
-      </SEO>
-      <Box sx={{ position: "relative", marginTop: 21 }}>
-        <Box
+      />
+      <Box sx={{ position: "relative", marginTop: 17 }}>
+        {/* <Box
+            sx={{
+              position: "absolute",
+              zIndex: -2,
+              bottom: '-220px',
+              left: '20%',
+              background: 'radial-gradient(closest-side at 50% 50%, #5189C8 0%, #5189C800 100%) ',
+              transform: 'matrix(-0.19, 0.98, -0.98, -0.19, 0, 0)',
+              width: 1609,
+              height: 1264,
+              pointerEvents: 'none',
+              opacity: 0.7,
+            }}
+          /> */}
+        {/* <Box
           ref={headerImage}
           sx={{
             position: "absolute",
@@ -232,404 +121,86 @@ const IndexPage = ({ location }: PageProps) => {
             transform: "translate(0px, -25%)",
             background: `url(${shape1})`,
             backgroundRepeat: "no-repeat",
-            backgroundPosition: "top -240px center",
+            backgroundPosition: "top -540px center",
             backgroundSize: "2070px 1155px",
             width: "100%",
             height: 1155,
           }}
-        />
+        /> */}
         <HeroText
-          maxWidth={600}
-          heroSpans={["The best of", "blockchains"]}
-          subtitleText="Move information and value anywhere."
-        />
-        <Box
-          sx={{
-            m: "auto",
-            maxWidth: { xs: 240, sm: 600 },
-            textAlign: "center",
-          }}
-        >
-          <Box
-            sx={{
-              width: "calc( 100% - 16px )",
-              mx: "auto",
-              mt: 15.5,
-              display: "flex",
-              flexWrap: "wrap",
-              justifyContent: "center",
-            }}
-          >
-            {tvl !== undefined && (
-              <Box
-                sx={{
-                  mt: 2,
-                  mx: 1,
-                  pt: { xs: 1, sm: 0 },
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-evenly",
-                  flexBasis: { xs: "100%", sm: "calc(33.33333% - 16px)" },
-                  borderTop: "1px solid white",
-                }}
-              >
-                <Typography sx={featuredNumber}>
-                  ${amountFormatter(tvl, 2)}
-                </Typography>
-                <Typography variant="body2">in TVL</Typography>
-              </Box>
-            )}
-            <Box
-              sx={{
-                mt: 2,
-                mx: 1,
-                pt: { xs: 1, sm: 0 },
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "space-evenly",
-                flexBasis: { xs: "100%", sm: "calc(33.33333% - 16px)" },
-                borderTop: "1px solid white",
-              }}
-            >
-              <Typography sx={featuredNumber}>{chainEnums.length}</Typography>
-              <Typography variant="body2">chain integrations</Typography>
-            </Box>
-            {messageTotal && (
-              <Box
-                sx={{
-                  mt: 2,
-                  mx: 1,
-                  pt: { xs: 1, sm: 0 },
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-evenly",
-                  flexBasis: { xs: "100%", sm: "calc(33.33333% - 16px)" },
-                  borderTop: "1px solid white",
-                }}
-              >
-                <Typography sx={featuredNumber}>
-                  {amountFormatter(messageTotal, 0)}
-                </Typography>
-                <Typography variant="body2">txs</Typography>
-              </Box>
-            )}
-          </Box>
-        </Box>
-      </Box>
-
-      <Box sx={{ position: "relative" }}>
-        <Box
-          ref={gradient1}
-          sx={{
-            position: "absolute",
-            zIndex: -1,
-            top: "50%",
-            background:
-              "radial-gradient(closest-side at 50% 50%, #E72850 0%, #E7285000 100%)",
-            transform: "matrix(0.96, 0.29, -0.29, 0.96, 0, 0)",
-            left: "60%",
-            width: 1645,
-            height: 903,
-            pointerEvents: "none",
-            opacity: 0.7,
-          }}
-        />
-        <Box
-          ref={gradient2}
-          sx={{
-            position: "absolute",
-            zIndex: -1,
-            top: "65%",
-            background:
-              "radial-gradient(closest-side at 50% 50%, #5189C8 0%, #5189C800 100%) ",
-            transform: "matrix(0.67, 0.74, -0.74, 0.67, 0, 0)",
-            left: "5%",
-            width: 1136,
-            height: 1489,
-            pointerEvents: "none",
-            opacity: 0.64,
-          }}
-        />
-        <Box
-          ref={shapeLeft}
-          sx={{
-            position: "absolute",
-            zIndex: -1,
-            background: `url(${shape})`,
-            backgroundSize: "contain",
-            top: -100,
-            right: "70vw",
-            width: 1363,
-            height: 1130,
-            pointerEvents: "none",
-            display: { xs: "none", md: "block" },
-          }}
-        />
-        <Box
-          ref={row1}
-          sx={{
-            display: "flex",
-            flexWrap: "wrap",
-            maxWidth: 1220,
-            px: 3.75,
-            mt: 35,
-            mx: "auto",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <Box
-            sx={{
-              flexBasis: { xs: "100%", md: "40%" },
-              flexGrow: 1,
-            }}
-          >
-            <Box sx={{ px: { xs: 0, md: 4 } }}>
-              <Box sx={{ maxWidth: 348, mx: "auto" }}>
-                <Typography variant="h3">
-                  <Box component="span" sx={{ color: "#FFCE00" }}>
-                    Protocol:{" "}
-                  </Box>
-                  <Box component="span" sx={{ display: "inline-block" }}>
-                    the core layer
-                  </Box>
-                </Typography>
-                <Typography sx={{ mt: 2 }}>
-                  The foundation that an ecosystem of apps is built on top of.
-                </Typography>
-                <Button
-                  component={RouterLink}
-                  to="/buidl/"
-                  sx={{ mt: 3 }}
-                  variant="outlined"
-                  color="inherit"
-                  endIcon={<ArrowForward />}
-                >
-                  Learn More
-                </Button>
-              </Box>
-            </Box>
-          </Box>
-          <Box
-            sx={{
-              mt: { xs: 8, md: null },
-              flexBasis: { xs: "100%", md: "60%" },
-              textAlign: "center",
-              flexGrow: 1,
-            }}
-          >
-            <Box
-              component="img"
-              src={protocols}
-              alt=""
-              sx={{ maxWidth: "100%" }}
-            />
-          </Box>
-        </Box>
-      </Box>
-
-      <Box
-        ref={row2}
-        sx={{
-          display: "flex",
-          flexWrap: "wrap-reverse",
-          maxWidth: 1220,
-          px: 3.75,
-          mt: 15.5,
-          mx: "auto",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        <Box
-          sx={{
-            mt: { xs: 8, md: null },
-            flexBasis: { xs: "100%", md: "60%" },
-            textAlign: "center",
-            flexGrow: 1,
-          }}
-        >
-          <Box component="img" src={apps} alt="" sx={{ maxWidth: "100%" }} />
-        </Box>
-        <Box sx={{ flexBasis: { xs: "100%", md: "40%" }, flexGrow: 1 }}>
-          <Box sx={{ px: { xs: 0, md: 4 } }}>
-            <Box sx={{ maxWidth: 348, mx: "auto" }}>
-              <Typography variant="h3">
-                <Box component="span" sx={{ color: "#FFCE00" }}>
-                  Apps:{" "}
-                </Box>
-                <Box component="span">endless possibilities</Box>
-              </Typography>
-              <Typography sx={{ mt: 2 }}>
-                Apps can now live across chains at once and integrate the best
-                of each.
-              </Typography>
-              <Button
-                component={RouterLink}
-                to="/apps/"
-                sx={{ mt: 3 }}
-                variant="outlined"
-                color="inherit"
-                endIcon={<ArrowForward />}
-              >
-                Learn More
-              </Button>
-            </Box>
-          </Box>
-        </Box>
-      </Box>
-
-      <Box sx={{ position: "relative" }}>
-        <Box
-          ref={shapeRight}
-          sx={{
-            position: "absolute",
-            zIndex: -1,
-            background: `url(${shape2})`,
-            backgroundSize: "contain",
-            top: "50%",
-            transform: "translateY(-50%)",
-            left: "75vw",
-            width: 1612,
-            height: 1316,
-            pointerEvents: "none",
-            display: { xs: "none", md: "block" },
-          }}
-        />
-        <Box
-          ref={row3}
-          sx={{
-            display: "flex",
-            flexWrap: "wrap",
-            maxWidth: 1220,
-            px: 3.75,
-            mt: 15.5,
-            mx: "auto",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <Box
-            sx={{
-              flexBasis: { xs: "100%", md: "40%" },
-              flexGrow: 1,
-            }}
-          >
-            <Box sx={{ px: { xs: 0, md: 4 } }}>
-              <Box sx={{ maxWidth: 340, mx: "auto" }}>
-                <Typography variant="h3">
-                  <Box component="span" sx={{ color: "#FFCE00" }}>
-                    Portal:{" "}
-                  </Box>
-                  <Box component="span" sx={{ display: "inline-block" }}>
-                    a token bridge
-                  </Box>
-                </Typography>
-                <Typography sx={{ mt: 2 }}>
-                  Never have to retrace your steps, with unlimited transfers
-                  across chains for tokens and NFTs wrapped by Wormhole.
-                </Typography>
-                <Button
-                  href={portalUrl}
-                  sx={{ mt: 3 }}
-                  variant="outlined"
-                  color="inherit"
-                  endIcon={<ArrowForward />}
-                >
-                  Learn More
-                </Button>
-              </Box>
-            </Box>
-          </Box>
-          <Box
-            sx={{
-              mt: { xs: 8, md: null },
-              flexBasis: { xs: "100%", md: "60%" },
-              textAlign: "center",
-              flexGrow: 1,
-            }}
-          >
-            <Button
-              variant="text"
-              href={portalUrl}
-              sx={{
-                "&:hover": {
-                  background: "transparent",
-                },
-                span: {
-                  display: "none",
-                },
-              }}
-            >
-              <Box
-                component="img"
-                src={portal}
-                alt=""
-                sx={{ maxWidth: "100%" }}
-              />
-            </Button>
-          </Box>
-        </Box>
-      </Box>
-
-      <Box ref={row4} sx={{ textAlign: "center", mt: 12.5, px: 2 }}>
-        <Typography variant="h3">
-          <Box component="span" sx={{ color: "#FFCE00" }}>
-            Cross-chain
-          </Box>
-          <Box component="span"> everything</Box>
-        </Typography>
-        <Typography sx={{ mt: 2, maxWidth: 480, mx: "auto", fontWeight: 300 }}>
-          Each blockchain has a distinct strength. Wormhole lets you get the
-          best out of every blockchain without compromise.
-        </Typography>
-      </Box>
-      <Box ref={row5} sx={{ maxWidth: 1220, mx: "auto", mt: 12, px: 3.75 }}>
-        <GridWithCards
-          data={[
-            {
-              src: cross,
-              size: 220,
-              header: "Never stop expanding",
-              description:
-                "Chains, information, and users are growing everyday. Build on a protocol that is set up to scale, with no limits, right from the start.",
-            },
-            {
-              src: blob,
-              size: 220,
-              header: "Explore and experiment",
-              description:
-                "Now is the time to explore and experiment. The only limit to what you're able to build is your imagination.",
-            },
-            {
-              src: cube,
-              size: 220,
-              header: "Power your project",
-              description:
-                "Join the growing list of projects that are composing, raising, and succeeding with Wormhole core layer.",
-            },
+          heroSpans={["Check the Stats"]}
+          subtitleText={[
+            "Explore real-time movement of information and",
+            "value around the Alephium bridge.",
           ]}
         />
+      </Box>
+      <Box sx={{position: 'relative'}}>
+      {/* <Box
+          ref={gradient_1}
+          sx={{
+            position: "absolute",
+            zIndex: -2,
+            top: '0',
+            background: 'radial-gradient(closest-side at 50% 50%, #5189C8 0%, #5189C800 100%) ',
+            transform: 'matrix(-0.67, 0.74, -0.74, -0.67, 0, 0)',
+            left: '70%',
+            width: 1077,
+            height: 1329,
+            pointerEvents: 'none',
+            opacity: 0.64,
+          }}
+        />  */}
+      {/* <Box
+            ref={shapeRight}
+            sx={{
+              position: "absolute",
+              zIndex: -1,
+              background: `url(${shapes})`,
+              backgroundSize: 'contain',
+              top: '0',
+              left: "85%",
+              transform: 'scaleX(-1)',
+              width: 1227,
+              height: 1018,
+              pointerEvents: 'none',
+              display:{xs: 'none', md: 'block'},
+            }}
+          />   */}
+      <Box sx={{ maxWidth: 1220, mx: "auto", mt: 30, px: 3.75 }}>
+        <Box
+          sx={{
+            px: 4,
+            display: "flex",
+            flexWrap: "wrap",
+            alignItems: "center",
+          }}
+        >
+          <Box sx={{ flexGrow: 1 }} />
+          <NetworkSelect />
+        </Box>
+      </Box>
+
+
+        <Box sx={{ maxWidth: 1220, mx: "auto", px: 3.75 }}>
+
+          {doneReadingQueryParams && <>
+
+            <ExplorerSearch location={location} />
+
+            {!(emitterChain && emitterAddress && targetChain && sequence) && // if there is no messageId query &&
+              !txId && (                                      // if there is no transactionId query
+                <ExplorerStats
+                  emitterChain={emitterChain}
+                  emitterAddress={emitterAddress}
+                />
+              )}
+
+          </>}
+        </Box>
       </Box>
     </Layout>
   );
 };
 
-type IndexQueryType = {
-  site: {
-    siteMetadata: {
-      siteUrl: string;
-    };
-  };
-};
-const IndexStaticQuery = graphql`
-  query Index {
-    site {
-      siteMetadata {
-        siteUrl
-      }
-    }
-  }
-`;
-
-export default IndexPage;
+export default ExplorerPage;
